requirements.txt
Biopython>=1.81
click>=8.1
pydantic>=2.6
requests>=2.31
intervaltree>=3.1

